% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_cat.R
\docType{data}
\name{SI_cat}
\alias{SI_cat}
\title{Example of 20 observations for categorical analysis}
\format{
A data frame with 20 rows and 8 variables:
\describe{
\item{study}{study ID}
\item{n1, n2, n3}{number of participants in each group}
\item{w1, w2, w3}{number of withdrawals in each group}
}
}
\usage{
SI_cat
}
\description{
Sample from Sato/Iwamoto dataset of 35 studies with withdrawal data\cr
see Bolland MJ, Gamble GD, Avenell A, Cooper DJ, Grey A. Participant withdrawals were unusually distributed in randomized trials with integrity concerns: a statistical investigation. J Clin Epidemiol 2021;131:22-29.
}
\keyword{datasets}
