% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_cohort.R
\docType{data}
\name{SI_cohort}
\alias{SI_cohort}
\title{Example of 100 observations for cohort analysis}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{study}{study ID}
\item{var}{variable name}
\item{n}{number of participants in group}
\item{m}{mean of variable}
\item{s}{sd of variable}
\item{group}{number of group}
}
}
\usage{
SI_cohort
}
\description{
Sample from Sato/Iwamoto dataset of 226 baseline variables in 34 cohorts\cr
see Bolland MJ, Gamble GD, Avenell A, Grey A. Identical summary statistics were uncommon in randomized trials and cohort studies. J Clin Epidemiol 2021;136:180-188.
}
\keyword{datasets}
